Move repo diff calculation and remove persistence
Things are simpler if the diffs are calcualted in PullRequestUpdater.
Also, there is no need for persistence, since at any point in time the
diffs should simply come from the source manifest on the target branch and
the source manifest on the head branch.
